用最暴力的方式实现最小二乘参数估计。flavor2,19阶。计算结果调试看正确

+++ ecs/matrix.h
@@ -7,7 +7,7 @@
 
 #include <cmath>
 
-bool Gauss(float *A, float *B, int n);  //求逆
+bool Gauss(double *A, double *B, int n);  //求逆
 bool matrix_mul(double* A, int A_rows, int A_cols, double* B, int B_rows, int B_cols, double *result);//矩阵乘法
 void matrix_trans(double *M, int M_rows, int M_cols,  double *m_trans);//转置
